clean: move more functions from main to helper.cpp
