+++ src/helper.cpp
@@ -1,6 +1,10 @@
 #include "globals.h"
+#include <cstring>
+#include <fstream>
 #include <iostream>
 #include <memory>
+#include <sstream>
+#include <string>
 
 const std::string token_to_string(const Token &tok) {
   switch (tok) {
@@ -224,4 +228,57 @@ print_lexed_tokens_test(std::unique_ptr<TokenChunk[]> &token_stack) {
     i++;
   }
   return ss.str();
+}
+
+std::string readFile(const std::string &filename) {
+  std::ifstream file(filename);
+  if (!file.is_open()) {
+    std::cerr << "Unable to open source file" << std::endl;
+  }
+  std::string source((std::istreambuf_iterator<char>(file)),
+                     std::istreambuf_iterator<char>());
+  file.close();
+  return source;
+}
+
+void writeFile(const std::string &filename, const std::string &output) {
+  std::ofstream file(filename);
+  if (!file.is_open()) {
+    std::cerr << "Unable to open output file" << std::endl;
+  }
+  file << output;
+  file.close();
+}
+
+bool get_flags(const int &argc, char *argv[], std::string &filename,
+               std::string &parseString, std::string &test_string) {
+  if (argc == 1)
+    return 0;
+  if (argc == 2) {
+    filename = argv[1];
+    return 0;
+  } else if (argc == 3) {
+    if (std::strcmp(argv[1], "-e")) {
+      parseString = argv[2];
+    }
+  }
+  int ind = 0;
+  char *currentFlag;
+  char *currentOption;
+  bool isTest{false};
+  for (int count{0}; count < argc; count++) {
+    if (count % 2) {
+      currentFlag = argv[count];
+    } else {
+      if (strcmp(currentFlag, "-e") == 0) {
+        parseString = argv[count];
+      } else if (strcmp(currentFlag, "-t") == 0) {
+        isTest = true;
+        test_string = argv[count];
+      } else {
+        filename = argv[count];
+      }
+    }
+  }
+  return isTest;
 }
